requirements.txt
python-telegram-bot==20.3
tensorflow==2.12.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
nltk==3.8.1
joblib==1.2.0
python-dotenv==1.0.0